pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "docx-json-replacer"
version = "0.5.1"
description = "Replace template text in DOCX files with values from JSON"
readme = "README.md"
authors = [{name = "liuspatt", email = "liuspatt@example.com"}]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
requires-python = ">=3.7"
dependencies = [
    "docxtpl>=0.20.0",
    "python-docx>=0.8.11",
]

[project.scripts]
docx-json-replacer = "docx_json_replacer.cli:main"

[project.urls]
Homepage = "https://github.com/liuspatt/docx-json-replacer"
Repository = "https://github.com/liuspatt/docx-json-replacer"
Issues = "https://github.com/liuspatt/docx-json-replacer/issues"